clean up AppState a bit
The original intention of initializing an AppState with database_connection being an Option
was so that we could call AppState::new prior to having a database connection.
I think this new way is fine as well since we are not accessing things like config.database_uri
directly outside of the apps top-level main function. This change allows us to remove a bunch of
unwrap()s